requirements.txt
asgiref==3.5.2
Django==4.0.5
django-countries==7.3.2
django-phonenumber-field==6.1.0
phonenumbers==8.12.50
pika==1.2.1
pyuca==1.2
sqlparse==0.4.2
typing-extensions==4.2.0
tzdata==2022.1
